Refactor: Update sync status handling and UI text formatting
This commit refines the synchronization process and improves how UI text, particularly for time differences, is handled.

Key changes:

- **Sync Status in Repository:**
    - `SyncRepositoryImpl` now sends `SyncStatus.SYNCING` only when the `WorkInfo.State` is `RUNNING`.
    - Removed redundant `SyncStatus.SYNCING` emission at the beginning of `performFullSync`.
- **Sync Status in ViewModel:**
    - `SettingsViewModel` now updates `isSyncing` to `true` *before* collecting the `performFullSync()` flow.
    - Corrected `isSyncing` to be set to `false` when `SyncStatus.FAILED` is received.
- **DateTime Formatting:**
    - `DateTimeExt.kt`: The `minutesAgo` value is now cast to `Int` when constructing the `UiText.StringResource` for `R.string.minutes_ago`.
- **UI Text Resource Handling:**
    - `UiText.kt`: `StringResource.asString()` (both composable and context versions) now correctly passes the `args` to `stringResource()` and `context.getString()` respectively, allowing for formatted strings with placeholders.

+++ app/src/main/java/com/fomaxtro/notemark/presentation/util/DateTimeExt.kt
@@ -49,7 +49,7 @@ fun Instant?.toSyncDateTimeUiText(): UiText {
                 else -> {
                     val minutesAgo = this.until(now, ChronoUnit.MINUTES)
 
-                    UiText.StringResource(R.string.minutes_ago, arrayOf(minutesAgo))
+                    UiText.StringResource(R.string.minutes_ago, arrayOf(minutesAgo.toInt()))
                 }
             }
         }
